requirements.txt
affine==2.4.0
APScheduler==3.10.0
attrs==22.2.0
certifi==2022.12.7
cftime==1.6.2
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
ecmwf-opendata==0.1.2
Fiona==1.9.1
idna==3.4
multiurl==0.2.0
munch==2.5.0
netCDF4==1.6.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rasterio==1.3.5.post1
requests==2.28.2
rioxarray==0.13.3
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
xarray==2023.2.0
xmltodict==0.13.0
